Retrieve build job ID number when queued
It looks like in the case of queuing a job, the Location returned by
POSTing to the corresponding job is not the actual build job ID number,
but an entry to the queue. This is further complicated by the fact that
the job is kept waiting for a few seconds (throttling?) before its ID
number is actually available.